+++ code/central/admin.py
@@ -15,13 +15,3 @@ admin.site.register(UserProfile, UserProfileAdmin)
 class OrganizationInvitationAdmin(admin.ModelAdmin):
     list_display = ("email_to_invite", "invited_by", "organization")
 admin.site.register(OrganizationInvitation, OrganizationInvitationAdmin)
-
-
-class FeedListingAdmin(admin.ModelAdmin):
-    list_display = ("title", "author", "posted_date")
-admin.site.register(FeedListing, FeedListingAdmin)
-
-
-class SubscriptionAdmin(admin.ModelAdmin):
-    list_display = ("email", "timestamp", "ip",)
-admin.site.register(Subscription, SubscriptionAdmin)
