+++ code/central/management/commands/package_for_download.py
@@ -27,14 +27,14 @@ from django.core.management import call_command
 from django.core.management.base import BaseCommand, CommandError
 
 import fle_utils.platforms
+from code.settings import LOG as logging
 from fle_utils.general import get_module_source_file
 from fle_utils.platforms import is_windows, system_script_extension, system_specific_zipping, system_specific_unzipping
 from kalite.management.commands.zip_kalite import create_default_archive_filename, Command as ZipCommand
-from kalite.settings import LOG as logging
+from kalite.updates.management.commands.update import Command as UpdateCommand
 from securesync import engine
 from securesync.management.commands.initdevice import Command as InitCommand
 from securesync.models import Zone, DeviceZone, Device, ChainOfTrust, ZoneInvitation
-from updates.management.commands.update import Command as UpdateCommand
 
 
 def install_from_package(data_json_file, signature_file, zip_file, dest_dir=None, install_files=[]):
