+++ code/central/models.py
@@ -184,20 +184,3 @@ class DeletionRecord(ExtendedModel):
     deleter = models.ForeignKey(User, related_name="deletion_actor")
     deleted_user = models.ForeignKey(User, related_name="deletion_recipient", blank=True, null=True)
     deleted_invite = models.ForeignKey(OrganizationInvitation, blank=True, null=True)
-
-
-class FeedListing(ExtendedModel):
-    title = models.CharField(max_length=150)
-    author = models.CharField(max_length=100)
-    description = models.TextField(blank=True)
-    posted_date = models.DateTimeField()
-    url = models.URLField()
-
-    def get_absolute_url(self):
-        return self.url
-
-class Subscription(ExtendedModel):
-    email = models.EmailField()
-    timestamp = models.DateTimeField(auto_now_add=True)
-    ip = models.CharField(max_length=100, blank=True)
-
