+++ code/faq/settings.py
@@ -1,0 +1,3 @@
+import os
+
+TEMPLATE_DIRS = (os.path.join(os.path.dirname(__file__), "templates"),)
