+++ ka-lite/python-packages/fle_utils/feeds/urls.py
@@ -1,0 +1,10 @@
+from django.conf.urls.defaults import patterns, include, url
+
+from . import RssSiteNewsFeed, AtomSiteNewsFeed
+
+
+urlpatterns = patterns('',
+    url(r'^addsubscription/$', 'add_subscription', {}, 'add_subscription'),
+    url(r'^feeds/rss/$', RssSiteNewsFeed(), {}, 'rss_feed'),
+    url(r'^feeds/atom/$', AtomSiteNewsFeed(), {}, 'atom_feed'),
+)
